Safari-609.1.11.1/LayoutTests/http/tests/media/video-served-as-text-with-type-expected.txt


EXPECTED (video.error == 'null') OK
EVENT(loadstart)
EVENT(loadedmetadata)
loaded media file served as text with type= attribute OK
END OF TEST
